requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==20.8
google-generativeai==0.3.2
groq==0.4.2
aiosqlite==0.19.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
cachetools==5.3.2
tenacity==8.2.3
